moved the folders to src
