+++ src/screens/EventAttendee/Register/styles.tsx
