+++ src/screens/EventOrganizer/Register/styles.tsx
