+++ src/screens/OtpVerify/styles.tsx
